feat: add backend report ability for Excel and Pdf, also add front progressbar, optimize return order steps and some other things...

+++ frontend/src/views/user/PageReturnOrderDetail.vue
@@ -1,9 +1,9 @@
 <template>
   <template v-if="!loading">
     <base-message
-        :has-close="false"
-        class="rounded-lg"
-        type="warning"
+      :has-close="false"
+      class="rounded-lg"
+      type="warning"
     >
       <div class="leading-relaxed">
         در صورتی می‌توانید اطلاعات سفارش مرجوع شده را تغییر دهید که توسط سایت تغییر وضعیت داده نشده باشد.
@@ -19,8 +19,8 @@
               <span class="tracking-widest">{{ returnOrder?.order_code }}</span>
             </div>
             <router-link
-                :to="{name: 'user.order.detail', params: {code: returnOrder?.order_code}}"
-                class="flex gap-2 items-center text-blue-600 group ml-auto mr-auto sm:ml-0"
+              :to="{name: 'user.order.detail', params: {code: returnOrder?.order_code}}"
+              class="flex gap-2 items-center text-blue-600 group ml-auto mr-auto sm:ml-0"
             >
               <span>مشاهده جزئیات</span>
               <ArrowLongLeftIcon class="w-5 h-5 group-hover:-translate-x-1.5 transition"/>
@@ -32,15 +32,15 @@
 
     <div class="mb-3">
       <partial-badge-status-return-order
-          :color-hex="returnOrder?.status.color_hex"
-          :text="returnOrder?.status.text"
-          class="w-full py-2 !text-sm"
+        :color-hex="returnOrder?.status.color_hex"
+        :text="returnOrder?.status.text"
+        class="w-full py-2 !text-sm"
       />
     </div>
   </template>
   <template v-else>
     <div
-        class="flex gap-2 items-center justify-between bg-white animate-pulse w-full h-8 rounded my-3 border-2 border-teal-500 mb-3 px-3">
+      class="flex gap-2 items-center justify-between bg-white animate-pulse w-full h-8 rounded my-3 border-2 border-teal-500 mb-3 px-3">
       <div class="w-40 h-3 rounded bg-slate-300"></div>
       <div class="w-16 h-3 rounded bg-blue-300"></div>
     </div>
@@ -58,24 +58,61 @@
       </h2>
 
       <base-loading-panel
-          :loading="loading"
-          type="form"
+        :loading="loading"
+        type="form"
       >
         <template #content>
           <partial-card class="border-0">
             <template #body>
               <div class="p-3">
-                <div
+                <div class="flex flex-wrap items-center gap-2 5">
+                  <div
                     v-if="!returnOrder?.has_status_changed"
                     class="mb-4 text-left"
-                >
-                  <base-button
-                      class="bg-rose-500 px-4 text-sm !py-1"
+                  >
+                    <base-button
+                      :disabled="cancelRequestLoading"
+                      class="!border-2 border-rose-500 bg-rose-500 px-4 text-sm !py-1"
                       type="button"
                       @click="handleCancelReturnOrder"
+                    >
+                      <VTransitionFade>
+                        <loader-circle
+                          v-if="cancelRequestLoading"
+                          big-circle-color="border-transparent"
+                          main-container-klass="absolute w-full h-full top-0 left-0"
+                          spinner-klass="!w-6 !h-6"
+                        />
+                      </VTransitionFade>
+
+                      <span>لغو درخواست</span>
+                    </base-button>
+                  </div>
+
+                  <div
+                    v-if="returnOrder?.next_status?.value"
+                    class="mb-4 text-left"
                   >
-                    لغو درخواست
-                  </base-button>
+                    <base-button
+                      :disabled="nextRequestStatusLoading"
+                      class="!border-2 border-black !text-black px-4 text-sm !py-1"
+                      type="button"
+                      @click="handleNextStatus"
+                    >
+                      <VTransitionFade>
+                        <loader-circle
+                          v-if="nextRequestStatusLoading"
+                          big-circle-color="border-transparent"
+                          container-bg-color="bg-black"
+                          main-container-klass="absolute w-full h-full top-0 left-0"
+                          small-circle-color="border-t-white"
+                          spinner-klass="!w-6 !h-6"
+                        />
+                      </VTransitionFade>
+
+                      <span>{{ returnOrder.next_status.text }}</span>
+                    </base-button>
+                  </div>
                 </div>
 
                 <div class="flex flex-wrap items-center gap-2 mb-3">
@@ -85,16 +122,16 @@
                 <div class="mb-3">
                   <base-textarea
                     :in-edit-mode="!returnOrder?.description || returnOrder?.description?.trim() === ''"
-                      :is-editable="!returnOrder?.has_status_changed"
-                      :value="returnOrder?.description"
-                      label-title="علت مرجوع نمودن محصول"
-                      name="description"
-                      placeholder="توضیحات خود را وارد نمایید..."
+                    :is-editable="!returnOrder?.has_status_changed"
+                    :value="returnOrder?.description"
+                    label-title="علت مرجوع نمودن محصول"
+                    name="description"
+                    placeholder="توضیحات خود را وارد نمایید..."
                   />
                 </div>
 
                 <template
-                    v-if="returnOrder?.not_accepted_description && returnOrder?.not_accepted_description?.trim() !== ''"
+                  v-if="returnOrder?.not_accepted_description && returnOrder?.not_accepted_description?.trim() !== ''"
                 >
                   <partial-input-label title="پاسخ به درخواست"/>
                   <div class="flex flex-col gap-3 shadow-md bg-indigo-50 p-6 mt-3 border border-slate-50 leading-loose">
@@ -116,10 +153,10 @@
       </h2>
 
       <base-message
-          v-if="!loading"
-          :has-close="false"
-          class="rounded-lg"
-          type="info"
+        v-if="!loading"
+        :has-close="false"
+        class="rounded-lg"
+        type="info"
       >
         <div class="leading-relaxed">
           برای عدم انتخاب کالا برای مرجوع، تعداد آن را
@@ -129,37 +166,37 @@
       </base-message>
 
       <base-loading-panel
-          :loading="loading"
-          type="list"
+        :loading="loading"
+        type="list"
       >
         <template #content>
           <partial-card class="border-0">
             <template #body>
               <ul
-                  v-if="items?.length"
-                  class="divide-y divide-gray-100"
+                v-if="items?.length"
+                class="divide-y divide-gray-100"
               >
                 <li
-                    v-for="product in items"
-                    :key="product.id"
-                    class="relative flex flex-col md:flex-row mb-2 py-3 pr-3 pl-10"
+                  v-for="product in items"
+                  :key="product.id"
+                  class="relative flex flex-col md:flex-row mb-2 py-3 pr-3 pl-10"
                 >
                   <div class="shrink-0">
                     <router-link
-                        v-if="product?.product"
-                        :to="{name: 'product.detail', params: {slug: product.product.slug}}"
-                        class="inline-block"
+                      v-if="product?.product"
+                      :to="{name: 'product.detail', params: {slug: product.product.slug}}"
+                      class="inline-block"
                     >
                       <base-lazy-image
-                          :alt="product.product.title"
-                          :lazy-src="product.image.path"
-                          :size="FileSizes.SMALL"
-                          class="!w-36 md:!w-24 h-auto hover:scale-95 transition"
+                        :alt="product.product.title"
+                        :lazy-src="product.image.path"
+                        :size="FileSizes.SMALL"
+                        class="!w-36 md:!w-24 h-auto hover:scale-95 transition"
                       />
                     </router-link>
                     <div
-                        v-else
-                        class="inline-block text-center size-36 border rounded md:size-24 hover:scale-95 transition"
+                      v-else
+                      class="inline-block text-center size-36 border rounded md:size-24 hover:scale-95 transition"
                     >
                       <div class="flex flex-col justify-center gap-2 h-full">
                         <PhotoIcon class="size-8 text-rose-300 mx-auto"/>
@@ -170,23 +207,23 @@
 
                   <div class="grow mr-3 mb-2 md:mb-0">
                     <router-link
-                        v-if="product?.product"
-                        :to="{name: 'product.detail', params: {slug: product.product.slug}}"
-                        class="inline-block mb-2 text-blue-600 hover:text-opacity-90 leading-relaxed"
+                      v-if="product?.product"
+                      :to="{name: 'product.detail', params: {slug: product.product.slug}}"
+                      class="inline-block mb-2 text-blue-600 hover:text-opacity-90 leading-relaxed"
                     >
                       {{ product.product_title }}
                     </router-link>
                     <div
-                        v-else
-                        class="inline-block mb-2 text-blue-600 hover:text-opacity-90 leading-relaxed"
+                      v-else
+                      class="inline-block mb-2 text-blue-600 hover:text-opacity-90 leading-relaxed"
                     >
                       {{ product.product_title }}
                     </div>
 
                     <div class="flex flex-wrap mb-2 items-center">
                       <div
-                          v-if="product.discounted_price && getPercentageOfPortion(product.discounted_price, product.price) > 0"
-                          class="rounded-lg bg-rose-500 text-white py-1 px-2 my-1 ml-3 flex items-center justify-center"
+                        v-if="product.discounted_price && getPercentageOfPortion(product.discounted_price, product.price) > 0"
+                        class="rounded-lg bg-rose-500 text-white py-1 px-2 my-1 ml-3 flex items-center justify-center"
                       >
                         <span class="text-xs">%</span>
                         <div class="mr-1 inline-block text-sm">
@@ -206,11 +243,11 @@
                         </div>
 
                         <div
-                            v-if="product.discounted_price"
-                            class="relative text-center"
+                          v-if="product.discounted_price"
+                          class="relative text-center"
                         >
                         <span
-                            class="absolute top-1/2 -translate-y-1/2 left-0 h-[1px] w-full bg-slate-500 -rotate-3"></span>
+                          class="absolute top-1/2 -translate-y-1/2 left-0 h-[1px] w-full bg-slate-500 -rotate-3"></span>
                           <div class="text-slate-500 text-sm">
                             {{ numberFormat(product.price) }}
                             <span class="text-xs text-gray-400">تومان</span>
@@ -223,9 +260,9 @@
                       <div class="flex flex-col gap-2 items-center justify-center">
                         <span class="text-pink-500 text-xs">تعداد محصول بازگشتی</span>
                         <base-spinner
-                            v-model="product.return_quantity"
-                            :max="product.quantity"
-                            :min="0"
+                          v-model="product.return_quantity"
+                          :max="product.quantity"
+                          :min="0"
                         />
                       </div>
                       <span v-if="product.is_accepted"
@@ -235,20 +272,20 @@
 
                   <div class="shrink-0 mr-3 text-sm md:w-72">
                     <div
-                        v-if="product.color_name"
-                        class="mb-2 flex items-center"
+                      v-if="product.color_name"
+                      class="mb-2 flex items-center"
                     >
                       <span class="text-gray-600 ml-2 text-xs">رنگ:</span>
                       {{ product.color_name }}
                       <span
-                          v-tooltip.top="'' + product.color_name + ''"
-                          :style="'background-color:' + product.color_hex"
-                          class="inline-block w-5 h-5 rounded-full border mr-2"
+                        v-tooltip.top="'' + product.color_name + ''"
+                        :style="'background-color:' + product.color_hex"
+                        class="inline-block w-5 h-5 rounded-full border mr-2"
                       ></span>
                     </div>
                     <div
-                        v-if="product.size"
-                        class="mb-2"
+                      v-if="product.size"
+                      class="mb-2"
                     >
                       <span class="text-gray-600 ml-2 text-xs">سایز:</span>
                       {{ product.size }}
@@ -262,8 +299,8 @@
               </ul>
 
               <div
-                  v-else
-                  class="text-orange-300 text-center p-2"
+                v-else
+                class="text-orange-300 text-center p-2"
               >
                 هیچ محصولی خریداری نشده است!
               </div>
@@ -274,19 +311,19 @@
     </div>
 
     <div
-        v-if="!loading"
-        class="mt-3"
+      v-if="!loading"
+      class="mt-3"
     >
       <base-animated-button
-          :disabled="!canSubmit"
-          class="bg-emerald-500 text-white mr-auto px-6 w-full sm:w-auto"
-          type="submit"
+        :disabled="!canSubmit"
+        class="bg-emerald-500 text-white mr-auto px-6 w-full sm:w-auto"
+        type="submit"
       >
         <VTransitionFade>
           <loader-circle
-              v-if="!canSubmit"
-              big-circle-color="border-transparent"
-              main-container-klass="absolute w-full h-full top-0 left-0"
+            v-if="!canSubmit"
+            big-circle-color="border-transparent"
+            main-container-klass="absolute w-full h-full top-0 left-0"
           />
         </VTransitionFade>
 
@@ -298,11 +335,11 @@
       </base-animated-button>
 
       <div
-          v-if="Object.keys(errors)?.length"
-          class="text-left"
+        v-if="Object.keys(errors)?.length"
+        class="text-left"
       >
         <div
-            class="w-full sm:w-auto sm:inline-block text-center text-sm border-2 border-rose-500 bg-rose-50 rounded-full py-1 px-3 mt-2"
+          class="w-full sm:w-auto sm:inline-block text-center text-sm border-2 border-rose-500 bg-rose-50 rounded-full py-1 px-3 mt-2"
         >
           (
           <span>{{ Object.keys(errors)?.length }}</span>
@@ -316,7 +353,7 @@
 
 <script setup>
 import {onMounted, ref} from "vue";
-import {CheckIcon, ArrowLongLeftIcon, PhotoIcon} from "@heroicons/vue/24/outline/index.js";
+import {ArrowLongLeftIcon, CheckIcon, PhotoIcon} from "@heroicons/vue/24/outline/index.js";
 import BaseMessage from "@/components/base/BaseMessage.vue";
 import BaseLoadingPanel from "@/components/base/BaseLoadingPanel.vue";
 import PartialCard from "@/components/partials/PartialCard.vue";
@@ -330,7 +367,7 @@ import LoaderCircle from "@/components/base/loader/LoaderCircle.vue";
 import BaseAnimatedButton from "@/components/base/BaseAnimatedButton.vue";
 import BaseSpinner from "@/components/base/BaseSpinner.vue";
 import {useFormSubmit} from "@/composables/form-submit.js";
-import {numberFormat, getPercentageOfPortion, getRouteParamByKey} from "@/composables/helper.js";
+import {getPercentageOfPortion, getRouteParamByKey, numberFormat} from "@/composables/helper.js";
 import {UserPanelReturnOrderAPI} from "@/service/APIUserPanel.js";
 import {useToast} from "vue-toastification";
 import {useConfirmToast} from "@/composables/toast-helper.js";
@@ -376,23 +413,58 @@ const {canSubmit, errors, onSubmit} = useFormSubmit({
   })
 })
 
-function handleCancelReturnOrder() {
-  useConfirmToast(() => {
-    UserPanelReturnOrderAPI.deleteById(codeParam.value, {
-      success() {
-        router.push({name: 'user.return_orders'})
-      },
-    })
-  }, 'انصراف از مرجوع کالا')
-}
-
-onMounted(() => {
+function fetchRequestedReturnOrder() {
   UserPanelReturnOrderAPI.fetchById(codeParam.value, {
     success(response) {
       setFormFields(response.data)
       loading.value = false
     },
   })
+}
+
+const cancelRequestLoading = ref(false)
+
+function handleCancelReturnOrder() {
+  if (cancelRequestLoading.value) return
+
+  useConfirmToast(() => {
+    cancelRequestLoading.value = true
+
+    UserPanelReturnOrderAPI.deleteById(codeParam.value, {
+      success() {
+        router.push({name: 'user.return_orders'})
+      },
+      finally() {
+        cancelRequestLoading.value = false
+      },
+    })
+  }, 'انصراف از مرجوع کالا')
+}
+
+const nextRequestStatusLoading = ref(false)
+
+function handleNextStatus() {
+  if (nextRequestStatusLoading.value || !returnOrder?.next_status?.value) return
+
+  useConfirmToast(() => {
+    nextRequestStatusLoading.value = true
+
+    UserPanelReturnOrderAPI.changeStatus(codeParam.value, {
+      status: returnOrder.next_status.value
+    }, {
+      success() {
+        toast.success('تغییر وضعیت با موفقیت انجام شد.')
+        fetchRequestedReturnOrder()
+      },
+      finally() {
+        nextRequestStatusLoading.value = false
+      },
+    })
+  }, 'تغییر وضعیت مرجوع', 'لطفا دقت نمایید، وضعیت مرجوع در حال تغییر به ' + '[' + returnOrder.next_status.text + ']' + ' می‌باشد.')
+}
+
+onMounted(() => {
+  fetchRequestedReturnOrder()
 })
 
 function setFormFields(item) {
